third_party/blink/web_tests/platform/generic/fast/dom/attribute-empty-value-no-children-expected.txt

Attr empty value tests Bug 16923

In Opera, Firefox and IE, when an Attr's value is an empty string, the Attr node won't have any childNodes. The following 12 tests will see if this is true for WebKit for both parsed and dynamically-created Attr nodes. The tests use multiple methods of changing the Attr's value and even test reverting from a non-empty value to an empty one to check that all childNodes were removed. Some of the tests make use of .textContent, so this test is not compatible with IE. Opera and Firefox both completely pass this test.

SubTest 1 = PASS
SubTest 2 = PASS
SubTest 3 = PASS
SubTest 4 = PASS
SubTest 5 = PASS
SubTest 6 = PASS
SubTest 7 = PASS
SubTest 8 = PASS
SubTest 9 = PASS
SubTest 10 = PASS
Complete Test = PASS
